// src/services/models/tradeTransactionPaginationResponse.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Swipe Pad API
 * OpenAPI spec version: 0.0.12
 */
import type { TradeTransactionResponse } from "./tradeTransactionResponse";

export interface TradeTransactionPaginationResponse {
  readonly total: number;
  readonly page: number;
  readonly pageSize: number;
  readonly data: readonly TradeTransactionResponse[];
}