// src/services/models/tradeTransactionResponse.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Swipe Pad API
 * OpenAPI spec version: 0.0.12
 */
import type { MemeResponse } from "./memeResponse";

export interface TradeTransactionResponse {
  readonly type: string;
  readonly hash: string;
  readonly blockNumber: number;
  readonly timestamp: number;
  readonly tokenAddress: string;
  readonly trader: string;
  readonly tokenAmount: string;
  readonly quoteAmount: string;
  readonly virtualTokenReserve: string;
  readonly virtualQuoteReserve: string;
  readonly priceAtTime: string;
  readonly priceAtTimeByUsd: string;
  readonly volumeAtTime: string;
  readonly volumeAtTimeByUsd: string;
  readonly version: string;
  readonly memeInfo: MemeResponse;
}